refactor: 태그 임시 데이터 파일 분리
