+++ src/data/tags.ts
@@ -1,0 +1,12 @@
+export const tags = [
+  { id: 1, name: "#ISFJ" },
+  { id: 2, name: "#Kpop" },
+  { id: 3, name: "#인스타그램" },
+  { id: 4, name: "#사진" },
+  { id: 5, name: "#야구" },
+  { id: 6, name: "#볼링" },
+  { id: 7, name: "#프로그래밍" },
+  { id: 8, name: "#게임" },
+  { id: 9, name: "#위스키" },
+  { id: 10, name: "#테마파크" },
+];
